+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+ENHANCEMENTS:
+* `r/tfe_workspace`: Add an `auto_destroy_activity_duration` attribute for automatic scheduling of auto-destroy runs based off of workspace activity, by @notchairmk [#1377](https://github.com/hashicorp/terraform-provider-tfe/pull/1377)
+* `d/tfe_workspace`: Add an `auto_destroy_activity_duration`, by @notchairmk [#1377](https://github.com/hashicorp/terraform-provider-tfe/pull/1377)
+
 ## v0.56.0
 
 ENHANCEMENTS: